Update train.py
update :  y_pred_ = tf.cast(y_pred > t, tf.int32)

+++ src/unetseg/train.py
@@ -262,7 +262,7 @@ def build_model_unetplusplus(cfg):
     def mean_iou(y_true, y_pred):
         prec = []
         for t in np.arange(0.5, 1.0, 0.05):
-            y_pred_ = tf.to_int32(y_pred > t)
+            y_pred_ = tf.cast(y_pred > t, tf.int32) # tf.to_int32(y_pred > t)
             score, up_opt = tf.metrics.mean_iou(y_true, y_pred_,cfg.n_classes)
             K.get_session().run(tf.local_variables_initializer())
             with tf.control_dependencies([up_opt]):
@@ -427,7 +427,7 @@ def build_model_unet(cfg):
     def mean_iou(y_true, y_pred):
         prec = []
         for t in np.arange(0.5, 1.0, 0.05):
-            y_pred_ = tf.to_int32(y_pred > t)
+            y_pred_ = tf.cast(y_pred > t, tf.int32) # tf.to_int32(y_pred > t)
             score, up_opt = tf.metrics.mean_iou(y_true, y_pred_,2)
             K.get_session().run(tf.local_variables_initializer())
             with tf.control_dependencies([up_opt]):
